Packages exception refactor
