+++ src/main/java/com/alkemy/ong/data/exceptions/ParamNotFoundException.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.data.exception;
+package com.alkemy.ong.data.exceptions;
 
 public class ParamNotFoundException extends RuntimeException{
     public ParamNotFoundException(String error){super(error);}
