+++ src/main/java/com/alkemy/ong/data/gateways/DefaultNewsGateway.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.data.gateways;
 
 import com.alkemy.ong.data.entities.NewsEntity;
-import com.alkemy.ong.data.exception.ParamNotFoundException;
+import com.alkemy.ong.data.exceptions.ParamNotFoundException;
 import com.alkemy.ong.data.repositories.NewsRepository;
 import com.alkemy.ong.domain.gateway.NewsGateway;
 import com.alkemy.ong.domain.model.News;
